remove bugs and nonsense
Normally I wouldn't change multiple things in a single commit,
but this is a tiny project for a school paper so I don't think it
matters too much.

For loop iterables were converted to tuples... don't ask me why
I tried to shoehorn 'del' before a function pops...
Defended my use of q instead of i for a for loop.
The more I look at it, the more I want to change, I'll cut my losses.

To be honest, this is bad code. Not only is it due for a rewrite, but
the goal of the project wasn't thought out very well. In my defense,
I made it in a frenzied couple of hours and at the very least, it has comments.